+++ src/components/button-minor/button-minor.component.tsx
@@ -2,12 +2,23 @@ import React from "react";
 import StyledButtonMinor from "./button-minor.style";
 import { ButtonProps } from "../button";
 
+export interface ButtonMinorProps extends ButtonProps {
+  /** @private @ignore */
+  isInPassword?: boolean;
+}
+
 export const ButtonMinor = ({
   buttonType = "secondary",
   size = "medium",
+  isInPassword,
   ...rest
-}: ButtonProps) => (
-  <StyledButtonMinor size={size} buttonType={buttonType} {...rest} />
+}: ButtonMinorProps) => (
+  <StyledButtonMinor
+    size={size}
+    buttonType={buttonType}
+    isInPassword={isInPassword}
+    {...rest}
+  />
 );
 
 ButtonMinor.displayName = "ButtonMinor";