+++ src/components/password/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./password.component";
+export type { PasswordProps } from "./password.component";